chore: Refactored DTO names to follow common convention
