+++ app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/FeaturesResponseDTO.java
@@ -1,4 +1,4 @@
-package com.appsmith.server.dtos.ce;
+package com.appsmith.server.dtos;
 
 import lombok.Getter;
 import lombok.Setter;
