+++ app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/ImportActionResultDTO.java
@@ -1,4 +1,4 @@
-package com.appsmith.server.dtos.ce;
+package com.appsmith.server.dtos;
 
 import com.appsmith.server.domains.NewAction;
 import lombok.Getter;
