+++ app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/ProductAlertMessageApplicabilityContext.java
@@ -1,4 +1,4 @@
-package com.appsmith.server.dtos.ce;
+package com.appsmith.server.dtos;
 
 public enum ProductAlertMessageApplicabilityContext {
     COMMON_CONFIG,
