+++ app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/ProductAlertResponseDTO.java
@@ -1,4 +1,4 @@
-package com.appsmith.server.dtos.ce;
+package com.appsmith.server.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
