+++ app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/RefactorEntityNameDTO.java
@@ -1,5 +1,5 @@
 package com.appsmith.server.dtos;
 
-import com.appsmith.server.dtos.ce.RefactorEntityNameDTO_CE;
+import com.appsmith.server.dtos.ce.RefactorEntityNameCE_DTO;
 
-public class RefactorEntityNameDTO extends RefactorEntityNameDTO_CE {}
+public class RefactorEntityNameDTO extends RefactorEntityNameCE_DTO {}
