+++ app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/UncommittedChangesDTO.java
@@ -1,4 +1,4 @@
-package com.appsmith.server.dtos.ce;
+package com.appsmith.server.dtos;
 
 import lombok.Data;
