+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ce/CacheableFeatureFlagHelperCE.java
@@ -1,8 +1,8 @@
 package com.appsmith.server.services.ce;
 
 import com.appsmith.server.domains.User;
-import com.appsmith.server.dtos.ce.FeaturesRequestDTO;
-import com.appsmith.server.dtos.ce.FeaturesResponseDTO;
+import com.appsmith.server.dtos.FeaturesRequestDTO;
+import com.appsmith.server.dtos.FeaturesResponseDTO;
 import com.appsmith.server.featureflags.CachedFeatures;
 import com.appsmith.server.featureflags.CachedFlags;
 import reactor.core.publisher.Mono;
