+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ce/ProductAlertServiceCE.java
@@ -1,6 +1,6 @@
 package com.appsmith.server.services.ce;
 
-import com.appsmith.server.dtos.ce.ProductAlertResponseDTO;
+import com.appsmith.server.dtos.ProductAlertResponseDTO;
 import reactor.core.publisher.Mono;
 
 import java.util.List;
